requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
streamlit==1.31.0

# LangChain and AI (Compatible versions)
langchain==0.3.13
langchain-google-genai==2.0.8
langchain-core==0.3.28
langchain-community==0.3.13
google-generativeai==0.8.3

# Git Operations
GitPython==3.1.41

# Code Parsing
esprima==4.0.1
javalang==0.13.0

# Data Validation (Compatible with LangChain)
pydantic==2.11.9
pydantic-settings==2.10.1

# Async Operations
aiofiles==23.2.1
python-multipart==0.0.6

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Streamlit Components
streamlit-mermaid==0.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0
